Move getMoveInterpretation() into boardmatcher module
