+++ modules/boardmatcher.js
@@ -1,6 +1,8 @@
 const {sgf} = require('./fileformats')
 const helper = require('./helper')
 
+exports.shapes = null
+
 exports.readShapes = function(filename) {
     let tree = sgf.parseFile(filename)[0]
     let result = []
@@ -60,17 +62,21 @@ exports.cornerMatch = function(points, target) {
     return i < 8 ? [i, false] : [i - 8, true]
 }
 
-exports.shapeMatch = function(shape, board, vertex, corner = false) {
+exports.shapeMatch = function(shape, board, vertex) {
     if (!board.hasVertex(vertex)) return null
+
     let sign = board.get(vertex)
     if (sign === 0) return null
 
+    let corner = 'type' in shape && shape.type === 'corner'
+
     for (let anchor of shape.candidates) {
-        let hypotheses = [...Array(8)].map(() => true)
+        let hypotheses = Array(8).fill(true)
+        let i = 0
 
         // Hypothesize vertex === anchor
 
-        if (corner && board.getSymmetries(anchor).every(([x, y]) => x !== vertex[0] || y !== vertex[1]))
+        if (corner && board.getSymmetries(anchor).every(v => !helper.vertexEquals(v, vertex)))
             continue
 
         for (let j = 0; j < shape.points.length; j++) {
@@ -86,12 +92,65 @@ exports.shapeMatch = function(shape, board, vertex, corner = false) {
                     hypotheses[k] = false
             }
 
-            if (!hypotheses.includes(true)) break
+            i = hypotheses.indexOf(true)
+            if (i < 0) break
         }
 
-        let symm = hypotheses.indexOf(true)
-        if (symm >= 0) return [symm, sign]
+        if (i >= 0) return [i, sign < 0]
     }
+
+    return null
+}
+
+exports.getMoveInterpretation = function(board, vertex, {shapes = null} = {}) {
+    if (!board.hasVertex(vertex)) return 'Pass'
+
+    if (shapes == null) {
+        if (exports.shapes == null) {
+            exports.shapes = exports.readShapes(`${__dirname}/../data/shapes.sgf`)
+        }
+
+        shapes = exports.shapes
+    }
+
+    let sign = board.get(vertex)
+    let neighbors = board.getNeighbors(vertex)
+
+    // Check atari
+
+    if (neighbors.some(v => board.get(v) === -sign && board.getLiberties(v).length === 1))
+        return 'Atari'
+
+    // Check connection
+
+    let friendly = neighbors.filter(v => board.get(v) === sign)
+    if (friendly.length === neighbors.length) return 'Fill'
+    if (friendly.length >= 2) return 'Connect'
+
+    // Match shape
+
+    for (let shape of shapes) {
+        if ('size' in shape && (board.width !== board.height || board.width !== +shape.size))
+            continue
+
+        if (exports.shapeMatch(shape, board, vertex))
+            return shape.name
+    }
+
+    if (friendly.length === 1) return 'Stretch'
+
+    // Determine position to edges
+
+    if (helper.vertexEquals(vertex, [(board.width - 1) / 2, (board.height - 1) / 2]))
+        return 'Tengen'
+
+    let diff = board.getCanonicalVertex(vertex).map(x => x + 1)
+
+    if (!helper.vertexEquals(diff, [4, 4]) && board.getHandicapPlacement(9).some(v => helper.vertexEquals(v, vertex)))
+        return 'Hoshi'
+
+    if (diff[1] <= 6)
+        return diff.join('-') + ' point'
 
     return null
 }
